% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_daisy.R
\name{geom_daisy}
\alias{geom_daisy}
\title{daisy}
\usage{
geom_daisy(x = 1, y = 1, petal = "pink", center = "yellow", scale = 1)
}
\arguments{
\item{x}{= 1 x position}

\item{y}{= 1 y position}

\item{petal}{petal colour}

\item{center}{center colour}

\item{scale}{scale of daisy}
}
\value{
ggplot daisy
}
\description{
\code{geom_daisy} creates a ggplot daisy.
}
\examples{

library(ggplot2)
library(ggforce)

ggplot() +
 geom_daisy() +
 coord_fixed() +
 theme_void()


}
